Playing with representation of floating part of numbers
+ Added ft_pow

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: mykman <mykman@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/10 15:46:33 by mykman            #+#    #+#             */
-/*   Updated: 2021/05/28 02:11:22 by mykman           ###   ########.fr       */
+/*   Updated: 2021/05/28 17:06:04 by mykman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -93,7 +93,7 @@ int				ft_lstsize(t_list *lst);
 
 int				ft_intsize(int n, int c);
 int				ft_index(const char *s, char c);
-int				ft_pow(int n, int pow);
+unsigned long	ft_pow(unsigned long n, int pow);
 void			ft_padding(char c, int n);
 char			*ft_ctoa(char c);
 char			*ft_ultoa_base(unsigned long n, int minsize, char *base);
